tests/wpt/tests/css/css-multicol/parsing/column-width-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Multi-column Layout: parsing column-width with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-multicol/#propdef-column-width">
<meta name="assert" content="column-width supports the full grammar 'auto | <length>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("column-width", "auto");

test_valid_value("column-width", "0", "0px");
test_valid_value("column-width", "10px");
</script>
</body>
</html>
